Function nano_core::nano_core_start

#[no_mangle]
pub extern "C" fn nano_core_start(
    multiboot_information_virtual_address: usize, 
    early_double_fault_stack_top: usize
)

The main entry point into Theseus, that is, the first Rust code that the Theseus kernel runs.

This is called from assembly code entry point for Theseus, found in nano_core/src/boot/arch_x86_64/boot.asm.

This function does the following things:

Bootstraps the OS, including logging and basic early exception handlers

Sets up basic virtual memory

Initializes the state_store module

Finally, calls the Captain module, which initializes and configures the rest of Theseus.

If a failure occurs and is propagated back up to this function, the OS is shut down.

Note

In general, you never need to modify the nano_core to change Theseus’s behavior, because the nano_core is essentially logic-agnostic boilerplate code and set up routines. If you want to customize or change how the OS components are initialized, then change the captain::init routine.